Add hook for custom/work alias dotfile
This commit is one of the first commits that will support extending
current dotfiles.

Prior to this commit, I created a private repository titled
"work-dotfiles" which I clone locally, and contains the
`.<dotfile>.local` files. This is so that I can have a more portable
setup on a per-environment basis.