fix: visual block working in galaxy line
I was getting an error in GalaxyLine when I was doing a visual block
highlight -- I got a cannot concatenate nil value on the guifg nvim
command.

After investigating, it turns out that the visual block highlight (which
shows up at ^V or ^I) is ACTUALLY intrepeted as \22, hence that
mysterious addition.